history summary fix
 - adjusted prompt
 - adjusted citation removal
 - length cutoff by words, not characters
