+++ backend/onyx/configs/agent_configs.py
@@ -11,7 +11,7 @@ AGENT_DEFAULT_NUM_DOCS_FOR_REFINED_DECOMPOSITION = 5
 AGENT_DEFAULT_EXPLORATORY_SEARCH_RESULTS = 3
 AGENT_DEFAULT_MIN_ORIG_QUESTION_DOCS = 3
 AGENT_DEFAULT_MAX_ANSWER_CONTEXT_DOCS = 10
-AGENT_DEFAULT_MAX_STATIC_HISTORY_CHAR_LENGTH = 10000
+AGENT_DEFAULT_MAX_STATIC_HISTORY_WORD_LENGTH = 2000
 
 #####
 # Agent Configs
@@ -72,9 +72,9 @@ AGENT_MAX_ANSWER_CONTEXT_DOCS = int(
 )  # 8
 
 
-AGENT_MAX_STATIC_HISTORY_CHAR_LENGTH = int(
-    os.environ.get("AGENT_MAX_STATIC_HISTORY_CHAR_LENGTH")
-    or AGENT_DEFAULT_MAX_STATIC_HISTORY_CHAR_LENGTH
-)  # 10000
+AGENT_MAX_STATIC_HISTORY_WORD_LENGTH = int(
+    os.environ.get("AGENT_MAX_STATIC_HISTORY_WORD_LENGTH")
+    or AGENT_DEFAULT_MAX_STATIC_HISTORY_WORD_LENGTH
+)  # 2000
 
 GRAPH_VERSION_NAME: str = "a"
